{\rtf1\ansi\ansicpg1252\cocoartf1038\cocoasubrtf250 {\fonttbl\f0\fswiss\fcharset0 Helvetica;} {\colortbl;\red255\green255\blue255;} \margl1440\margr1440\vieww9000\viewh8400\viewkind0 \pard\tx720\tx1440\tx2160\tx2880\tx3600\tx4320\tx5040\tx5760\tx6480\tx7200\tx7920\tx8640\ql\qnatural\pardirnatural
\f0\fs24 \cf0 This revision is absolutely positively unstable, DO NOT USE IT! \
\
Removed unnecessary files and folders.\
\
Removed ProMOL_302. If you want the latest stable go to revision 57.\
\
Changed amino acid pictures names so that they are not confusing.\
\
Changes to ProMol.py include:\
	separating the code into multiple files\
	Optimizing the code\
	standardizing the formatting\
This is the first commit in several to come that will dramatically change the code. The functionality in most cases will not be changed.\
\
Much of the indentation is not correct due an attempt to standardize. Indentations will be fixed in a future revision.}