Changes to ConSCRIPT to make it compatible with Linux.
Many changes to ProMol.
Changed motif maker to have it's own tab, and am still working on making it compatible with motif finder.
Made changes to make it more compatible with linux and need to make more changes in that area.
Changed the GUI in many areas.
Deleted many motifs that don't seem to have a use right now.